+++ src/main/java/platformer/database/bridge/SQLStorage.java
@@ -1,7 +1,7 @@
 package platformer.database.bridge;
 
 import platformer.core.Account;
-import platformer.database.BoardItem;
+import platformer.model.BoardItem;
 import platformer.database.Settings;
 import platformer.debug.logger.Logger;
 import platformer.debug.logger.Message;
