+++ src/main/java/platformer/model/BoardItem.java
@@ -1,4 +1,4 @@
-package platformer.database;
+package platformer.model;
 
 public class BoardItem {
